fix: use is_void() instead of self_closing for HTML closing tags
Non-void elements like textarea must always emit a closing tag regardless
of self-closing syntax in source. Moved is_void + VOID_ELEMENTS to
svelte_ast (shared layer) and updated codegen to check element name
instead of parser flag. Renamed test case to non_void_self_closing.
