+++ tasks/compiler_tests/cases2/mixed_html_elements/case-rust.js
@@ -1,5 +1,5 @@
 import * as $ from "svelte/internal/client";
-var root = $.from_html(`<input/> <textarea/> <area/> <br/> <a></a>`, 1);
+var root = $.from_html(`<input/> <textarea></textarea> <area/> <br/> <a></a>`, 1);
 export default function App($$anchor) {
 	var fragment = root();
 	$.next(8);
