+++ tasks/compiler_tests/cases2/non_void_self_closing/case-svelte.js
@@ -1,5 +1,5 @@
 import * as $ from "svelte/internal/client";
-var root = $.from_html(`<textarea/>`);
+var root = $.from_html(`<textarea></textarea>`);
 export default function App($$anchor) {
 	var textarea = root();
 	$.append($$anchor, textarea);
