+++ tasks/compiler_tests/cases2/non_void_self_closing/case.svelte
